These 2 files are remnants of attempting the earlier planned method of renaming the .ssc files to .cs.  They are being deleted because my porting work will be done completely on my local disc, until such time that I want to merge ported code into the project.
With regards to the renaming problem, Stephan Tobies has found that the logs are kept on the CodePlex server, and are just not provided to the SVN client.  He has created a Codeplex work item at http://codeplex.codeplex.com/workitem/25490.  If it gets voted up sufficiently, it may get CodePlex to provide that log information, and thus help with future work regarding the Boogie depot.