Make the modelExceptions option an integer.
0 means no modeling at all.

1 (which can be specified only if also specifying whole program translation)
means that a fixpoint analysis will be done on all of the translated assemblies
to compute the set of exceptions that each method can throw (even though we're
using only the fact that a method throws anything at all, not the specific
exceptions it can throw) so only calls to methods that can throw an exception
have a branch after the call that checks to see if an exception has been
thrown. Methods that are defined outside of the set of translated assemblies
are assumed to *not* throw any exceptions!

2 means that every method is considered to be one that can throw an exception.