Added /useBaseNameForFile command line argument. The Scanner and Parser constructors have been modified to take an optional argument specifying this and the ExecutionEngine passes for that value CommandLineOptions.Clo.UseBaseNameForFileName
This option when true causes the basename of file to be used inside
created Tokens instead of what the user passed on the command line
which might be a relative or absolute path.

The motivation for adding this option is that it is needed for the
lit driven tests so that the output of Boogie can be reliably checked.