Fix nasty bug introduced by commit 61a94f409975.
There were many calls in the code to

``` new TokenTextWriter("<buffer>", buffer, false) ```

Prior to 61a94f409975 this was a call to the following constructor

```
TokenTextWriter(string filename, TextWriter writer, bool setTokens)
```

After that commit these then became calls to

```
TokenTextWriter(string filename, TextWriter writer, bool pretty)
```

An example of where this would cause issues was if ToString() was called on an
AbsyCmd then the its token would be modified because the setTokens parameter
was effectively set to True when the original intention was for it to be set
to false!

To fix this I've

* Removed the default parameter value for pretty and fixed all
uses so that we pass false where it was implicitly being set before

* Where the intention was to set setTokens to false this has been fixed so
  it is actually set to false!

Unfortunately I couldn't find a way of observing this bug from the Boogie
executable so I couldn't create a test case. I could only observe it when I was
using Boogie's APIs.