Fix bug in Duplicator where GotoCmd's LabelTargets and LabelNames where not updated during duplication to point to duplicated BasicBlocks.
Because the lists weren't being duplicated and resolved to the new basic blocks
a duplicated Implementation pointed into Blocks in the old implementation via
GotoCmds. This is bad and is not the expected behaviour.

Now if VisitImplementation() is called during duplication GotoCmds will be
resolved to point to duplicated Blocks.