Change the return type of StandardVisitor.VisitLiteralExpr() from LiteralExpr to Expr. Enforcing the return type be LiteralExpr is too restrictive.
There is precedence for this. For example VisitNAryExpr() returns
an Expr not an NAryExpr.

Unfortunately this a breaking API change so anyone who subclasses
the StandardVisitor (or one of its sub classes) and overrides this
method will get compilation errors until they change the return type.
a visitor that does constant folding of an Expr tree