Don't force the exit code to 0 by default
Although it's convenient to force the exit code to zero for lit testing,
it is more conventional (and better for interfacing with other systems)
to return a non-zero error if verification fail: at the moment, the only
way (IIUC) to find out whether verification succeeded (on the command
line) was to parse the textual output of Boogie.

This patch adds a CountVerificationErrors option, and enables it by
default. The patch also sets this option to zero for tests, lest lit get
confused in test cases where verification actually fails.