Use the EndCurly token when creating the ReturnCmd for unifiedExit
When there are more than one exit blocks, an unified exit block is
created which includes a ReturnCmd. However, the ReturnCmd is created
with NoToken. This causes the line/column number reported for a failed
postcondition to be (0,0). The right token should be the EndCurly since
the ReturnCmd is in the exit block.