[HACK] Fix #47 where Z3 4.5 would not work with Boogie. It seems the that the behaviour of
```
(set-option :TIMEOUT 0)
```

has changed. Previously it meant no timeout and now apparently
it means give no time for solving which seems weird...

For now implement some hacks to only set this when the value
is greater than zero.

The code here is a complete mess the timeout was being set in at least
three different places. I've removed one of the places because there
doesn't seem to be a good reason to immediatly send an the option to
Z3 in `SMTLibProcessTheoremProver.SetTimeOut(int)`. I've left the other
places intact for now.