Updated floating-point constant syntax
This implements the changes to the floating-point syntax proposed by
@RustanLeino in the following comments:

https://github.com/boogie-org/boogie/pull/91#issuecomment-378798275
https://github.com/boogie-org/boogie/pull/91#issuecomment-380968607

Read the linked comments for full details on what
has been changed. To clarify some points:

-The +zero and -zero special values have been removed
-The significand may be un-normalized; this allows the same number to be
written in multiple ways. For example, the number 10 may be written as
0xA0.0e-1f24e8, 0xA.0e0f24e8, 0x0.Ae1f24e8, 0x0.0Ae2f24e8, etc. (However, the
+/-infinity and NaN special values can only be written in a single way:
0+oo[sigSize]e[expSize], 0-oo[sigSize]e[expSize], and 0NaN[sigSize]e[expSize],
respectively)
-The significand must have trailing zeros such that the last nibble is fully
included. For example, in order to represent a floating-point value that has a
24-bit significand with the bit pattern 1_0000_0000_0000_0000_0000_001
(including the hidden bit at the beginning) and an exponent of 0, it may be
written as 0x1.000002e0f24e8, but not 0x1.000001e0f24e8.

The original floating-point syntax is specified in this wiki page:

https://github.com/boogie-org/boogie/wiki/Draft-floating-point-Boogie-language-extension