Introduce a user-callback for prover errors
In some cases, the prover was repoting an error. Boogie prints the error
to console, but continues execution. The next query to the prover then
hangs because the prover is in a broken state. This callback will allow
Boogie clients to take custom actions on a prover error. No-op by
default.