Added useArrayTheory to a SMACK regression
The regression was failing (Boogie would report an assertion violation)
when I switched to Z3 4.8.8. This is clearly due to incompleteness
of Z3 when array theory is not used, and so I enabled it now.
I have to question the purpose of this regression now, but I guess
we should just leave it in there.