Call WriteErrorInformation and the error reporter delegate consistently
There are three calls to `WriteErrorInformation` in this ExecutionEngine.cs.
Only of them were unconditional; the third one was conditional on `er` being
`null`.  Because of this inconsistency Dafny's LSP server was not reporting timeouts.

The original call to the printer was introduced in f1d7e9e9 (“Improve output for
diagnosing timeouts”).