Move interactive-only bits into subclass
This commit moves the portions of SMTLibProcessTheoremProver that are
only meaningful for interactive communication into a subclass, in
preparation for another subclass devoted to batch interaction.

It also pulls some internal portions of interactive-centric methods out
into separate, shared methods in the base class. These changes mostly
involve separating the steps of a) reading a response from the solver
and b) parsing the response. Batch solver interaction will do (a)
differently but can share code for (b).