Add more structured descriptions of proof obligations (#505)
Adds a structured type for describing proof obligations with potentially different text for success and failure forms. As a result, describing successful proofs becomes less awkward. This is more valuable for Boogie clients (such as Dafny) than it is for Boogie itself, though it has the benefit of grouping error messages in one place.

At the moment, Boogie only uses the failure text, but I could imagine it using the success text at some point. Dafny plans to use the success text right away, and to provide a large number of additional subclasses to attach to the instances of AssertCmd it generates.

I also hope that it could serve as a more general replacement for ErrorData, though I don't have a very thorough idea of how the various Boogie clients currently use that property.