Add `/randomSeedIterations` flag for verification stability analysis (#567)
This PR adds a new flag, `/randomSeedIterations:<n>`, which instructs Boogie to send each verification condition to the SMT solver `n` times, with the goal of analyzing the stability of proof across those `n` iterations. When used with `/vcsCores:<n>`, the iterations can be evaluated in parallel.

From the user perspective, the output of `/trace` and `/xml` now includes iteration numbers along with assertion batch indices.

Fixes #494.