Updated Z3 to version 4.11.2 (#672)
* Updated Z3 to version 4.11.2

* Ignoring else branches in the model

I believe that else branches are typically mostly (completely?)
unconstrained in the model. This leads to changes in else branch
values when Z3 is for example updated. This fix ignores checking
of else branches in the model and it should make this regression
more robust.

* Making regression more robust by excluding unconstrained model lines

* Making regression more robust by excluding unconstrained model lines

* Removed precise error trace matching

I don't think that is needed for this regression.

* Made keep going splits regression more robust

Introduced more properties to check to increase the needed SMT
resources. Properties are the same across two procedures now,
except that one does not allow splitting. Lowered SMT
resource count.

Co-authored-by: Shaz Qadeer <shaz.qadeer@gmail.com>