Fix to Lheap_Add translation (#740)
Lheap_Add primitive adds a cell to a linear heap. This PR changes the
translation so that only the domain of the target heap is modified. The
heap contents is assumed (magically) to contain the value being inserted
at the fresh address. This is sound because Ref T, the address type for
Lheap T is uninterpreted and potentially infinite source of cells with a
particular value.