Implement verification coverage analysis (#741)
This generalizes the previous feature of tracking "necessary
assumptions".

That feature, previously enabled with `-printNecessaryAssumptions`,
allowed each `assume` statement to be annotated with an `{:id ...}`
attribute, and Boogie would then report which assumptions were necessary
to complete the verification by requesting an unsat core from Z3.

Now that attribute can be placed on many other program elements:
* `assume` statements
* `assert` statements
* assignments
* calls
* requires clauses
* ensures clauses

Each of these is then labeled in the SMT encoding and will potentially
show up in the unsat core.

The new `-printVerificationCoverage` option (which replaces the old one)
prints the members of this broader set of program elements that show up
in the unsat core. With `-trace`, it also prints the coverage
information for each procedure as it finishes. The coverage information
for each procedure is also available in the `VCResult` produced at the
end of each verification.

Fixes #730.