Add missing antecedents to function axioms (#749)
Add type-constraining antecedents for each argument of a function when
generating the type axiom for that function.

These seem to have been intentionally but incorrectly left out in the
original implementation.

This fixes the missing antecedent issue in #735 and
dafny-lang/dafny#4053 but does not yet fix the missing type constructor
parameters. I have yet to track down exactly why those are left out.