Fix spec for allocation in Treiber stack (#751)
This PR makes the following changes:

1. Collector for Lheap instances now returns an empty set. This change
implies that the domain of a linear heap of type T no longer contributes
to the disjointness assumptions for the permission set Ref T. This
change is needed to allow valid references to heap objects to be wrapped
in an Lval. The new design appears to be reasonable because the
disjointness of heap domains is important only for heap erasure and does
not seem to be needed for property verification. On the other hand, it
seems important to be able to conveniently encode unique pointers to
heap addresses.

2. Lheap_Add is renamed to Lheap_Alloc and its output is wrapped inside
Lval. This change allows us to express the specification that successive
calls to allocations of Treiber stack return distinct references.

3. A new primitive Ref_Alloc is introduces to allocate a Ref T by itself
(i.e., without being associated with a Lheap instance). This is useful
to write the specification for the operation that allocates a Treiber
stack.