Allow absolute paths to prover plugins (#764)
Previously, the `/proverDll` option allowed only relative paths, since
it looked for `/` or `\` _within_ a name, but not at the beginning, to
determine whether the given name was a filename or a portion of an
internal DLL name. This changes it to also interpret a name as a
filename when either of those characters is the first character.