Print definition axioms for functions and constants. (#765)
When a function / constant has a definition axiom, in generated Boogie
files, these axioms were not printed inside `uses` clauses of that
function / constant. They were instead printed as top level
declarations. This caused issues when using that printed file with the
"/prune" option since some of these axioms could now be pruned.

This PR attempts to fix this issue such that axioms that define a
constant / function are now (only) printed inside `uses` clauses of the
function / constant they are defining.