Update version, and remove race condition (#811)
`ClearLocalSMTOptions()` was called _after_ the checker was released,
which leads to a race condition because a new thread may already be
using the checker's options while they're still being cleared. Here's an
example:
https://github.com/dafny-lang/dafny/actions/runs/6930078076/job/18849083894?pr=4773

This PR reduces statefulness of the code, since it no longer clears and
reparses `SMTLibProcessTheoremProver.options`. Instead it has a field
`IEnumerable<OptionValue> additionalSmtOptions` that can be used to
tweak SmtOptions between runs.