[Civl] Enhance quantifier elimination (#829)
TryElimination looks for a substitution for the lhs of single-static
assignments on a path-by-path basis through an atomic action. If an
assignment is of the form ```x := v``` where both x and v are variables,
it is possible that x is defined but v is not. In such a case, the code
change interprets the assignment as an equality and infers a definition
```v <-- x```.