[Civl] Combine linear heaps and maps into a single abstraction (#830)
This PR unifies Lmap and Lheap into a single abstraction with the latter
being defined in terms of the former.
```
type Lheap T = Lmap (Ref T) T
```
The type ```Ref T``` is explained below.

The implementation of Lmap primitives is done by modeling them directly
as pure procedures in Boogie syntax. Thus a lot of rewriting code in
LinearRewriter.cs is eliminated. However, inlining of atomic actions had
to be adjusted to allow for calling and inlining pure procedures from
atomic actions. It is possible that these changes can be reverted by
modeling these primitives as pure actions (which are already callable
from both yield procedures and actions).

An important innovation is separating the Lmap primitives from the
dynamic allocation of keys for which two separate types are provided.

```
type Loc _;
datatype Ref<V> {
  Ref(loc: Loc V),
  Nil()
}
```
The type ```Loc``` models a dynamically-allocated location. The
allocation itself is performed by the ```Loc_New``` primitive:
```
pure procedure {:inline 1} Loc_New<V>() returns (k: Lval (Loc V)) { }
```
The caller gets a fresh location (notice the ```Lval``` wrapper) every
time ```Loc_New``` is called. An instance of ```Loc T``` may be wrapped
into an instance of ```Ref T``` via the ```Ref``` constructor.

Two separate types ```Loc``` and ```Ref``` are needed because we want
the domains of different ```Lheap T``` values to be disjoint. At the
same time, we want to provide the capability to have unique references
to locations. If we unified ```Loc``` and ```Ref```, then we cannot have
both capabilities.