Allow `/prune:0` to disable pruning (#848)
Add a `/noprune` flag to turn off pruning. It's off by default, but this
can be convenient if a) it's turned on earlier on the command line, or
b) Boogie is being used from another system like Dafny.

I would have made it `/prune:{1,0}`, but it's harder to make that
backward compatible with the current behavior.