[Civl] New design of permissions (#851)
This PR implements a new design for the permission system in Civl. There
are two important changes:
- It is no longer possible to specify custom permission collectors.
Instead there is a fixed type-directed algorithm for collecting
permissions.
- String domain names are eliminated. Instead, each domain is identified
by a permission type, which is the type of each element of the
permission sets obtained via permission collection. Any type may be used
as a permission type.
 
Collection is orchestrated via three basic types---One, Set, Map. Types
embedded with a datatype are collected via a recursive traversal. The
primitives for permission transfer are designed afresh. Their signatures
can be seen in base.bpl. These primitives are the only mechanism for
permission redistribution apart from passing permissions to pending
asyncs. Therefore, for atomic actions that do not create pending asyncs,
the need for verifying the correctness of permission redistribution is
eliminated.

The bulk of the change in this PR is updating the regression tests to
account for the new method for specifying and collecting permissions.

In addition to the changes mentioned above:
- bug fixes in monomorphization and quantifier instantiation
- reorganization of tests into folders; new folders civl/paxos and
civl/large-samples; folded civl/podelski into civl/samples.

Outstanding issues:
- Paxos proof is flaky; current suspicion is that there might be a bug
in pool-based quantifier instantiation for lambdas
- Tests for the linear type checker are far from adequate

---------

Co-authored-by: Shaz Qadeer <shaz@meta.com>