[Civl] Fix ChangRoberts (#858)
This PR provides a new proof for the Chang Roberts protocol for leader
election. The new proof is different from the existing Civl proof in two
ways:
- Stronger specification is proved. The new proof proves that a unique
leader will be elected in all terminal states. The old proof only proved
that conflicting leaders are not elected.
- The modeling of the protocol avoids the use of channels and achieves
all communication asynchronous calls.

---------

Co-authored-by: Shaz Qadeer <shaz@meta.com>