Add an experimental backend that uses Lean to discharge goals (#850)
This experimental new backend translates passive Boogie programs into
Lean programs including theorems that can be automatically discharged by
the [`lean-auto`](https://github.com/leanprover-community/lean-auto)
package.

**Note:** The code is almost entirely separate from the existing Boogie
codebase, and modifies that code in only tiny ways (to provide an
additional command and invoke the new backend when it's enabled).
Because of this, and because of its current experimental nature, I'm
happy to do as little or as much code review on the new directory as
other folks want.

It has several goals:
* To allow axioms written in Boogie to be proven to have models and
therefore to be consistent.
* To allow pre-processing goals before sending them to SMT, possibly
discharging some without needing SMT.
* To allow experimentation with new tactics and decision procedures that
will more frequently discharge goals without the need for SMT.
* To allow the use of SMT proof certificates to increase assurance in
solver results (though this will require additions to `lean-auto`).
* Possibly: to allow interactive proofs of certain goals, which could
use infrastructure such as
[`mathlib4`](https://github.com/leanprover-community/mathlib4).

At the moment, completing proofs with the backend is manual. When given
the `/printLean:file.lean` option, it prints the Lean program to a file right
before the verification starts. It will probably need to stay this way until we
build some additional Lean infrastructure. Ultimately, however, the plan is
to make it as integrated and automatic as the current SMT-based
verification pipeline.

It includes a new test workflow with test files in `Test/lean-auto`.
This currently checks that it succeeds on several of the tests in
`Test/textbook` but will gradually expand to cover more of the test
suite.