[Civl] Improvements to pool-based quantifier instantiation (#862)
- Added a pool hint to nondeterministic output parameter of One_New.
Since this procedure is polymorphic, the pool name "One_New" may
correspond to many types. Therefore, pool-based quantifier instantiation
is updated so that it can handle the same pool name attached to bound
variables of several types. The main idea is to be permissive in the
construction of instantiations but prune out illegal instances just
before the instantiation.

- Changed AtomicAllocNode in treiber-stack.bpl so that the output
variable ```success``` is explicitly initialized. This prevents
quantification over the initial (nondeterministic) value of
```success``` in the construction of the gate of AtomicAllocNode.
Ultimately, this results in better behaved VCs.

Together the two changes above reduce flakiness in the proof of
treiber-stack.bpl.

---------

Co-authored-by: Shaz Qadeer <shaz@meta.com>