Allow both timeLimit and rlimit together
Sometimes we want to set a resource limit _and_ a time limit, where the
goal of the former is to have deterministic bounds but the goal of the
latter is to ensure timely CI completion. In this case, a good value for
the time limit should be much larger than the time we would expect to be
required to exceed the resource limit. This will ensure that execution
is usually deterministic, but in the rare case where resource count and
time are far from each other it'll still guarantee reasonable
termination.