Parse model values of the form 5e1 (no decimal) (#861)
Previously, models that contained real literals with no decimal point
(e.g., 5e1) would lead to an exception in the model parser. This PR
adapts Boogie to recognize such values.

Ensuring that Z3 will emit one of these values in a model is tricky, and
a test for it would be fragile. But I've seen it appear in practice.