[Civl] Add support for preconditions to actions (#872)
This PR adds preconditions to actions for the purpose of providing extra
contextual information in left mover checks for inductive
sequentialization. These preconditions are proved to be
interference-free just like other preconditions. Their initiation is
established by the IS invariant during the step condition check for the
action being moved left. As a result of this enhancement, the use of
abstractions (and the eliminates clause) is eliminated from Civl
benchmarks, the strengthening achieved by the extra gates of the
abstractions now achieved by the preconditions.

The intention is to eliminate abstractions and the eliminates clause
from Civl. But that will be done in a separate PR.

---------

Co-authored-by: Shaz Qadeer <shaz@meta.com>