Add `push`, `pop`, `hide` and `reveal` commands (#891)
### Changes
- Add hide and reveal commands, that works together with pruning to
enable preventing specific edges in the dependency graph from being
traversed, thereby causing more to be pruned. An axiom can be marked as
`hideable`. The hide and reveal commands can hide or reveal a set of
functions, a state which is maintained throughout the execution of a
Boogie implementation. During pruning, if a live but hidden function
uses a hideable axiom, then the axiom will not be made live.
- Add the `pushScope` and `popScope` commands, which represents entering
and leaving lexical scopes. Right now, these commands only influence the
behavior of `hide` and `reveal`, but in the future they may also allow
local variable shadowing.

### Testing
- Add test `Reveal.bpl` which tests the new feature
- Add test `AssumeFalseSplit.bpl`, which tests an existing piece of the
behavior of splits