[Civl] Resolve constructed expressions in IS implementation (#901)
All generated expressions and code in the implementation of IS proof
rules must be resolved and typechecked. In particular, any expression
used later for substitution in another expression should be resolved and
typechecked immediately. If there is a typed expression E and a
substitution v -> X is applied to E to get E' where X is not
typechecked, then typechecking E' later might not do the job. This is
because typing proceeds bottom up and if a top level expression already
has a type, the type may not descend to the leaves.

Co-authored-by: Shaz Qadeer <shaz@meta.com>