Resolve Boogie test deadlocks (#932)
### Changes
- Made a change to the Dispose method of
`CustomStackSizePoolTaskScheduler.cs`, so it now interrupts all its
thread instead of joining them. I think this was causing the entire
Boogie process to hang after it was finished, causing tests to deadlock.
- Add `blame-hang-timeout` to the `dotnet test` invocation that runs the
NUnit tests, which allows identifying which test fails when the unit
tests timeout.
- Add an undocumented option `/processTimeLimit` that can be used by
tests to let Boogie stop itself after a certain timespan, allowing
developers to get a stacktrace of where Boogie was at the point of
timeout.
- Remove APIs for cancelling requests, which were only available
programmatically and not used by Dafny.
- Fix a bug that could cause a concurrency related exception in
`QuantifierInstantiationInfo`, which was introduced by
https://github.com/boogie-org/boogie/pull/862.
- Use an invariant culture somewhere in printing debug output, so the
Boogie tests locally pass on my machine, despite my European culture.
- Some renaming of variations of VerifyImplementation, to make the
differences more clear.
- Configure a 30 seconds timelimit per VC for all Boogie tests by
default. Added an exception for a few slow tests
- Let Boogie cancel checking of a VC after its timelimit has passed,
regardless of what the solver is doing

### Testing
- Did not add any additional tests
- Let the test-suite succeed without retries 4 times, in an attempt to
determine that it resolves the test instabilities we've had.